Refactor : move methods to count collection to CountableCollectionTrait.
